feat: dry handleAddToCartSubmit method in ProductActions ProductCard and SingleProduct component and add it to utils
